LayoutTests/imported/w3c/web-platform-tests/navigation-api/precommit-handler/precommitHandler-replace-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Should never run

Harness Error (FAIL), message = Unhandled rejection

FAIL precommitHandler for a replace navigation, reject before commit assert_unreached: committed must not fulfill Reached unreachable code
PASS precommitHandler for a replace navigation, reject after commit
PASS precommitHandler for a replace navigation, success
